cdk_import/requirements-test.txt
attrs==25.3.0
aws-cdk.alexa-ask==1.204.0
aws-cdk.assets==1.204.0
aws-cdk.aws-accessanalyzer==1.204.0
aws-cdk.aws-acmpca==1.204.0
aws-cdk.aws-amazonmq==1.204.0
aws-cdk.aws-amplify==1.204.0
aws-cdk.aws-amplifyuibuilder==1.204.0
aws-cdk.aws-apigateway==1.204.0
aws-cdk.aws-apigatewayv2==1.204.0
aws-cdk.aws-appconfig==1.204.0
aws-cdk.aws-appflow==1.204.0
aws-cdk.aws-appintegrations==1.204.0
aws-cdk.aws-applicationautoscaling==1.204.0
aws-cdk.aws-applicationinsights==1.204.0
aws-cdk.aws-appmesh==1.204.0
aws-cdk.aws-apprunner==1.204.0
aws-cdk.aws-appstream==1.204.0
aws-cdk.aws-appsync==1.204.0
aws-cdk.aws-aps==1.204.0
aws-cdk.aws-athena==1.204.0
aws-cdk.aws-auditmanager==1.204.0
aws-cdk.aws-autoscaling==1.204.0
aws-cdk.aws-autoscaling-common==1.204.0
aws-cdk.aws-autoscaling-hooktargets==1.204.0
aws-cdk.aws-autoscalingplans==1.204.0
aws-cdk.aws-backup==1.204.0
aws-cdk.aws-backupgateway==1.204.0
aws-cdk.aws-batch==1.204.0
aws-cdk.aws-billingconductor==1.204.0
aws-cdk.aws-budgets==1.204.0
aws-cdk.aws-cassandra==1.204.0
aws-cdk.aws-ce==1.204.0
aws-cdk.aws-certificatemanager==1.204.0
aws-cdk.aws-chatbot==1.204.0
aws-cdk.aws-cloud9==1.204.0
aws-cdk.aws-cloudformation==1.204.0
aws-cdk.aws-cloudfront==1.204.0
aws-cdk.aws-cloudtrail==1.204.0
aws-cdk.aws-cloudwatch==1.204.0
aws-cdk.aws-codeartifact==1.204.0
aws-cdk.aws-codebuild==1.204.0
aws-cdk.aws-codecommit==1.204.0
aws-cdk.aws-codedeploy==1.204.0
aws-cdk.aws-codeguruprofiler==1.204.0
aws-cdk.aws-codegurureviewer==1.204.0
aws-cdk.aws-codepipeline==1.204.0
aws-cdk.aws-codestar==1.204.0
aws-cdk.aws-codestarconnections==1.204.0
aws-cdk.aws-codestarnotifications==1.204.0
aws-cdk.aws-cognito==1.204.0
aws-cdk.aws-comprehend==1.204.0
aws-cdk.aws-config==1.204.0
aws-cdk.aws-connect==1.204.0
aws-cdk.aws-connectcampaigns==1.204.0
aws-cdk.aws-controltower==1.204.0
aws-cdk.aws-cur==1.204.0
aws-cdk.aws-customerprofiles==1.204.0
aws-cdk.aws-databrew==1.204.0
aws-cdk.aws-datapipeline==1.204.0
aws-cdk.aws-datasync==1.204.0
aws-cdk.aws-dax==1.204.0
aws-cdk.aws-detective==1.204.0
aws-cdk.aws-devopsguru==1.204.0
aws-cdk.aws-directoryservice==1.204.0
aws-cdk.aws-dlm==1.204.0
aws-cdk.aws-dms==1.204.0
aws-cdk.aws-docdb==1.204.0
aws-cdk.aws-docdbelastic==1.204.0
aws-cdk.aws-dynamodb==1.204.0
aws-cdk.aws-ec2==1.204.0
aws-cdk.aws-ecr==1.204.0
aws-cdk.aws-ecr-assets==1.204.0
aws-cdk.aws-ecs==1.204.0
aws-cdk.aws-efs==1.204.0
aws-cdk.aws-eks==1.204.0
aws-cdk.aws-elasticache==1.204.0
aws-cdk.aws-elasticbeanstalk==1.204.0
aws-cdk.aws-elasticloadbalancing==1.204.0
aws-cdk.aws-elasticloadbalancingv2==1.204.0
aws-cdk.aws-elasticsearch==1.204.0
aws-cdk.aws-emr==1.204.0
aws-cdk.aws-emrcontainers==1.204.0
aws-cdk.aws-emrserverless==1.204.0
aws-cdk.aws-events==1.204.0
aws-cdk.aws-eventschemas==1.204.0
aws-cdk.aws-evidently==1.204.0
aws-cdk.aws-finspace==1.204.0
aws-cdk.aws-fis==1.204.0
aws-cdk.aws-fms==1.204.0
aws-cdk.aws-forecast==1.204.0
aws-cdk.aws-frauddetector==1.204.0
aws-cdk.aws-fsx==1.204.0
aws-cdk.aws-gamelift==1.204.0
aws-cdk.aws-globalaccelerator==1.204.0
aws-cdk.aws-glue==1.204.0
aws-cdk.aws-grafana==1.204.0
aws-cdk.aws-greengrass==1.204.0
aws-cdk.aws-greengrassv2==1.204.0
aws-cdk.aws-groundstation==1.204.0
aws-cdk.aws-guardduty==1.204.0
aws-cdk.aws-healthlake==1.204.0
aws-cdk.aws-iam==1.204.0
aws-cdk.aws-identitystore==1.204.0
aws-cdk.aws-imagebuilder==1.204.0
aws-cdk.aws-inspector==1.204.0
aws-cdk.aws-inspectorv2==1.204.0
aws-cdk.aws-internetmonitor==1.204.0
aws-cdk.aws-iot==1.204.0
aws-cdk.aws-iot1click==1.204.0
aws-cdk.aws-iotanalytics==1.204.0
aws-cdk.aws-iotcoredeviceadvisor==1.204.0
aws-cdk.aws-iotevents==1.204.0
aws-cdk.aws-iotfleethub==1.204.0
aws-cdk.aws-iotfleetwise==1.204.0
aws-cdk.aws-iotsitewise==1.204.0
aws-cdk.aws-iotthingsgraph==1.204.0
aws-cdk.aws-iottwinmaker==1.204.0
aws-cdk.aws-iotwireless==1.204.0
aws-cdk.aws-ivs==1.204.0
aws-cdk.aws-ivschat==1.204.0
aws-cdk.aws-kafkaconnect==1.204.0
aws-cdk.aws-kendra==1.204.0
aws-cdk.aws-kendraranking==1.204.0
aws-cdk.aws-kinesis==1.204.0
aws-cdk.aws-kinesisanalytics==1.204.0
aws-cdk.aws-kinesisanalyticsv2==1.204.0
aws-cdk.aws-kinesisfirehose==1.204.0
aws-cdk.aws-kinesisvideo==1.204.0
aws-cdk.aws-kms==1.204.0
aws-cdk.aws-lakeformation==1.204.0
aws-cdk.aws-lambda==1.204.0
aws-cdk.aws-lambda-nodejs==1.204.0
aws-cdk.aws-lex==1.204.0
aws-cdk.aws-licensemanager==1.204.0
aws-cdk.aws-lightsail==1.204.0
aws-cdk.aws-location==1.204.0
aws-cdk.aws-logs==1.204.0
aws-cdk.aws-lookoutequipment==1.204.0
aws-cdk.aws-lookoutmetrics==1.204.0
aws-cdk.aws-lookoutvision==1.204.0
aws-cdk.aws-m2==1.204.0
aws-cdk.aws-macie==1.204.0
aws-cdk.aws-managedblockchain==1.204.0
aws-cdk.aws-mediaconnect==1.204.0
aws-cdk.aws-mediaconvert==1.204.0
aws-cdk.aws-medialive==1.204.0
aws-cdk.aws-mediapackage==1.204.0
aws-cdk.aws-mediastore==1.204.0
aws-cdk.aws-mediatailor==1.204.0
aws-cdk.aws-memorydb==1.204.0
aws-cdk.aws-msk==1.204.0
aws-cdk.aws-mwaa==1.204.0
aws-cdk.aws-neptune==1.204.0
aws-cdk.aws-networkfirewall==1.204.0
aws-cdk.aws-networkmanager==1.204.0
aws-cdk.aws-nimblestudio==1.204.0
aws-cdk.aws-oam==1.204.0
aws-cdk.aws-omics==1.204.0
aws-cdk.aws-opensearchserverless==1.204.0
aws-cdk.aws-opensearchservice==1.204.0
aws-cdk.aws-opsworks==1.204.0
aws-cdk.aws-opsworkscm==1.204.0
aws-cdk.aws-organizations==1.204.0
aws-cdk.aws-osis==1.204.0
aws-cdk.aws-panorama==1.204.0
aws-cdk.aws-personalize==1.204.0
aws-cdk.aws-pinpoint==1.204.0
aws-cdk.aws-pinpointemail==1.204.0
aws-cdk.aws-pipes==1.204.0
aws-cdk.aws-proton==1.204.0
aws-cdk.aws-qldb==1.204.0
aws-cdk.aws-quicksight==1.204.0
aws-cdk.aws-ram==1.204.0
aws-cdk.aws-rds==1.204.0
aws-cdk.aws-redshift==1.204.0
aws-cdk.aws-redshiftserverless==1.204.0
aws-cdk.aws-refactorspaces==1.204.0
aws-cdk.aws-rekognition==1.204.0
aws-cdk.aws-resiliencehub==1.204.0
aws-cdk.aws-resourceexplorer2==1.204.0
aws-cdk.aws-resourcegroups==1.204.0
aws-cdk.aws-robomaker==1.204.0
aws-cdk.aws-rolesanywhere==1.204.0
aws-cdk.aws-route53==1.204.0
aws-cdk.aws-route53-targets==1.204.0
aws-cdk.aws-route53recoverycontrol==1.204.0
aws-cdk.aws-route53recoveryreadiness==1.204.0
aws-cdk.aws-route53resolver==1.204.0
aws-cdk.aws-rum==1.204.0
aws-cdk.aws-s3==1.204.0
aws-cdk.aws-s3-assets==1.204.0
aws-cdk.aws-s3objectlambda==1.204.0
aws-cdk.aws-s3outposts==1.204.0
aws-cdk.aws-sagemaker==1.204.0
aws-cdk.aws-sam==1.204.0
aws-cdk.aws-scheduler==1.204.0
aws-cdk.aws-sdb==1.204.0
aws-cdk.aws-secretsmanager==1.204.0
aws-cdk.aws-securityhub==1.204.0
aws-cdk.aws-servicecatalog==1.204.0
aws-cdk.aws-servicecatalogappregistry==1.204.0
aws-cdk.aws-servicediscovery==1.204.0
aws-cdk.aws-ses==1.204.0
aws-cdk.aws-shield==1.204.0
aws-cdk.aws-signer==1.204.0
aws-cdk.aws-simspaceweaver==1.204.0
aws-cdk.aws-sns==1.204.0
aws-cdk.aws-sns-subscriptions==1.204.0
aws-cdk.aws-sqs==1.204.0
aws-cdk.aws-ssm==1.204.0
aws-cdk.aws-ssmcontacts==1.204.0
aws-cdk.aws-ssmincidents==1.204.0
aws-cdk.aws-sso==1.204.0
aws-cdk.aws-stepfunctions==1.204.0
aws-cdk.aws-supportapp==1.204.0
aws-cdk.aws-synthetics==1.204.0
aws-cdk.aws-systemsmanagersap==1.204.0
aws-cdk.aws-timestream==1.204.0
aws-cdk.aws-transfer==1.204.0
aws-cdk.aws-voiceid==1.204.0
aws-cdk.aws-vpclattice==1.204.0
aws-cdk.aws-waf==1.204.0
aws-cdk.aws-wafregional==1.204.0
aws-cdk.aws-wafv2==1.204.0
aws-cdk.aws-wisdom==1.204.0
aws-cdk.aws-workspaces==1.204.0
aws-cdk.aws-xray==1.204.0
aws-cdk.cloud-assembly-schema==1.204.0
aws-cdk.cloudformation-include==1.204.0
aws-cdk.core==1.204.0
aws-cdk.custom-resources==1.204.0
aws-cdk.cx-api==1.204.0
aws-cdk.lambda-layer-awscli==1.204.0
aws-cdk.lambda-layer-kubectl==1.204.0
aws-cdk.lambda-layer-node-proxy-agent==1.204.0
aws-cdk.region-info==1.204.0
cattrs==25.1.1
constructs==3.4.344
exceptiongroup==1.3.0
importlib_resources==6.5.2
jsii==1.113.0
publication==0.0.3
python-dateutil==2.9.0.post0
six==1.17.0
typeguard==2.13.3
typing_extensions==4.15.0
